refactoring styles and structure to be more atomic
